move test to PopoverAutoAddTest
